finish backend

+++ backend/controllers/product.controller.js
@@ -39,7 +39,7 @@ export const updateProduct = async (req, res) => {
   const { id } = req.params;
   const product = req.body;
 
-  if (mongoose.Types.ObjectId.isValid(id)) {
+  if (!mongoose.Types.ObjectId.isValid(id)) {
     return res.status(404).json({
       success: false,
       message: "Product not found for update invalid Id",
@@ -64,6 +64,14 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   // this id comes from the dynamic section of URI
   const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({
+      success: false,
+      message: "Product not found for update invalid Id",
+    });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({
@@ -72,6 +80,6 @@ export const deleteProduct = async (req, res) => {
     });
   } catch (error) {
     console.error(error.message);
-    res.status(404).json({ success: false, message: "Server Error" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
